This commit brings fundamental performance improvements to the core routines of florence, in particular low level dispatchers for the most complex material models implemented based on Fastor SIMD optimised routines. Removes the hidden performance bottleneck in applying Dirichlet boundary conditions (this step should be instantanuous now). SIMD optimised compiled way for computing kinematic measures and for computing geometric stiffnesses in displacement and displacement potential based formulations

+++ Florence/FunctionSpace/GetBases.py
@@ -22,7 +22,7 @@ def GetBases(C, Quadrature, info, useLagrange=False):
     ns=[]; Basis=[]; gBasisx=[]; gBasisy=[]
     if info=='tri':
         p=C+1
-        ns = (p+1)*(p+2)/2
+        ns = int((p+1)*(p+2)/2)
         Basis = np.zeros((ns,w.shape[0]),dtype=np.float64)
         gBasisx = np.zeros((ns,w.shape[0]),dtype=np.float64)
         gBasisy = np.zeros((ns,w.shape[0]),dtype=np.float64)
